Update function_storage.py
Added a lot more functions. Essentialyl focusing around the json file. There's now initial encryption and master_pw initialization

And then we also have a way to decrypt the json file if we know the master password. 

Now the functions can be used in the main.py to set up the logic and error handling